refactor(presets): extract scroll logic from outline body (#7862)
Move scroll to block logic and highlight block mask from `outline-panel-body` to another files so that it can be reuse by `outline-viewer`
