+++ packages/presets/src/fragments/outline/body/outline-panel-body.ts
@@ -1,4 +1,3 @@
-import type { EditorHost } from '@blocksuite/block-std';
 import type {
   EdgelessRootBlockComponent,
   NoteBlockModel,
@@ -14,14 +13,14 @@ import { LitElement, type PropertyValues, css, html, nothing } from 'lit';
 import { customElement, property, query, state } from 'lit/decorators.js';
 import { repeat } from 'lit/directives/repeat.js';
 
+import type { AffineEditorContainer } from '../../../editors/editor-container.js';
 import type {
   ClickBlockEvent,
   DisplayModeChangeEvent,
   FitViewEvent,
   SelectEvent,
-} from '../card/outline-card.js';
+} from '../utils/custom-events.js';
 
-import { getDocTitleByEditorHost } from '../../doc-title/doc-title.js';
 import '../card/outline-card.js';
 import { startDragging } from '../utils/drag.js';
 import {
@@ -29,6 +28,7 @@ import {
   getNotesFromDoc,
   isHeadingBlock,
 } from '../utils/query.js';
+import { highlightBlock, scrollToBlock } from '../utils/scroll.js';
 import './outline-notice.js';
 
 type OutlineNoteItem = {
@@ -114,11 +114,9 @@ export class OutlinePanelBody extends SignalWatcher(
 ) {
   private _changedFlag = false;
 
+  private _clearHighlightMask: (() => void) | null = null;
+
   private _docDisposables: DisposableGroup | null = null;
-
-  private _highlightMask: HTMLDivElement | null = null;
-
-  private _highlightTimeoutId: ReturnType<typeof setTimeout> | null = null;
 
   private _indicatorTranslateY = 0;
 
@@ -162,8 +160,7 @@ export class OutlinePanelBody extends SignalWatcher(
                 .number=${idx + 1}
                 .index=${note.index}
                 .doc=${this.doc}
-                .editorMode=${this.mode}
-                .editorHost=${this.editorHost}
+                .editorMode=${this.editor.mode}
                 .activeHeadingId=${this.activeHeadingId.value}
                 .status=${selectedNotesSet.has(note.note.id)
                   ? this._dragging
@@ -208,15 +205,6 @@ export class OutlinePanelBody extends SignalWatcher(
   private _clearDocDisposables() {
     this._docDisposables?.dispose();
     this._docDisposables = null;
-  }
-
-  private _clearHighlightMask() {
-    this._highlightMask?.remove();
-    this._highlightMask = null;
-    if (this._highlightTimeoutId) {
-      clearTimeout(this._highlightTimeoutId);
-      this._highlightTimeoutId = null;
-    }
   }
 
   /*
@@ -365,7 +353,7 @@ export class OutlinePanelBody extends SignalWatcher(
   }
 
   private _isEdgelessMode() {
-    return this.mode === 'edgeless';
+    return this.editor.mode === 'edgeless';
   }
 
   private _moveNotes(
@@ -413,80 +401,19 @@ export class OutlinePanelBody extends SignalWatcher(
       .enableNotesSorting=${false}
       .showPreviewIcon=${this.showPreviewIcon}
       @click=${() => {
-        if (this._isEdgelessMode() || !this.editorHost) return;
-
-        const docTitle = getDocTitleByEditorHost(this.editorHost);
-        if (!docTitle) return;
-
-        docTitle.scrollIntoView({
-          behavior: 'instant',
-          block: 'start',
-        });
+        if (!this.doc.root) return;
+        scrollToBlock(this.editor, this.doc.root.id);
+        this.activeHeadingId.value = this.doc.root.id;
       }}
     ></affine-outline-block-preview>`;
   }
 
   private _scrollToBlock(e: ClickBlockEvent) {
-    if (this._isEdgelessMode() || !this.editorHost) return;
-
-    const rootComponent = this.editorHost.querySelector('affine-page-root');
-    if (!rootComponent) return;
-
-    const { blockPath } = e.detail;
-    const path = [rootComponent.model.id, ...blockPath];
-    const block = this.editorHost.view.viewFromPath('block', path);
-    if (!block) return;
-
-    block.scrollIntoView({
-      behavior: 'instant',
-      block: 'center',
-      inline: 'center',
-    });
+    scrollToBlock(this.editor, e.detail.blockId);
 
     requestAnimationFrame(() => {
-      if (!rootComponent.viewport) {
-        console.error('viewport should exist');
-        return;
-      }
-
-      const blockRect = block.getBoundingClientRect();
-      const { top, left, width, height } = blockRect;
-
-      const {
-        top: offsetY,
-        left: offsetX,
-        scrollTop,
-        scrollLeft,
-      } = rootComponent.viewport;
-
-      if (!this._highlightMask) {
-        this._highlightMask = document.createElement('div');
-        rootComponent.append(this._highlightMask);
-      }
-
-      Object.assign(this._highlightMask.style, {
-        position: 'absolute',
-        top: `${top - offsetY + scrollTop}px`,
-        left: `${left - offsetX + scrollLeft}px`,
-        width: `${width}px`,
-        height: `${height}px`,
-        background: 'var(--affine-hover-color)',
-        borderRadius: '4px',
-        display: 'block',
-      });
-
-      this.activeHeadingId.value = block.model.id;
-
-      // Clear the previous timeout if it exists
-      if (this._highlightTimeoutId !== null) {
-        clearTimeout(this._highlightTimeoutId);
-      }
-
-      this._highlightTimeoutId = setTimeout(() => {
-        if (this._highlightMask) {
-          this._highlightMask.style.display = 'none';
-        }
-      }, 1000);
+      this._clearHighlightMask = highlightBlock(this.editor, e.detail.blockId);
+      this.activeHeadingId.value = e.detail.blockId;
     });
   }
 
@@ -659,7 +586,7 @@ export class OutlinePanelBody extends SignalWatcher(
     }
 
     this._clearDocDisposables();
-    this._clearHighlightMask();
+    this._clearHighlightMask?.();
   }
 
   override firstUpdated(): void {
@@ -697,7 +624,7 @@ export class OutlinePanelBody extends SignalWatcher(
       this.edgeless &&
       this._isEdgelessMode()
     ) {
-      this._clearHighlightMask();
+      this._clearHighlightMask?.();
       if (_changedProperties.get('mode') === undefined) return;
 
       requestAnimationFrame(() => this._zoomToFit());
@@ -743,7 +670,7 @@ export class OutlinePanelBody extends SignalWatcher(
   accessor edgeless!: EdgelessRootBlockComponent | null;
 
   @property({ attribute: false })
-  accessor editorHost!: EditorHost;
+  accessor editor!: AffineEditorContainer;
 
   @property({ attribute: false })
   accessor enableNotesSorting!: boolean;
@@ -753,9 +680,6 @@ export class OutlinePanelBody extends SignalWatcher(
 
   @property({ attribute: false })
   accessor insertIndex: number | undefined = undefined;
-
-  @property({ attribute: false })
-  accessor mode: 'page' | 'edgeless' = 'page';
 
   @property({ attribute: false })
   accessor noticeVisible!: boolean;
