+++ packages/presets/src/fragments/outline/outline-panel.ts
@@ -11,7 +11,7 @@ import './body/outline-notice.js';
 import './body/outline-panel-body.js';
 import { type OutlineSettingsDataType, outlineSettingsKey } from './config.js';
 import './header/outline-panel-header.js';
-import { observeActiveHeading } from './utils/heading-highlight.js';
+import { observeActiveHeading } from './utils/scroll.js';
 
 const styles = css`
   :host {
@@ -163,7 +163,7 @@ export class OutlinePanel extends SignalWatcher(WithDisposable(LitElement)) {
           .doc=${this.doc}
           .fitPadding=${this.fitPadding}
           .edgeless=${this.edgeless}
-          .editorHost=${this.host}
+          .editor=${this.editor}
           .activeHeadingId=${this._activeHeadingId$}
           .mode=${this.mode}
           .showPreviewIcon=${this._showPreviewIcon}
